data: use int64 instead of int for sizes

+++ chronica/internal/data/dataset.go
@@ -45,9 +45,9 @@ type Dataset interface {
 }
 
 // New creates a new dataset with the given arguments.
-func New[T btree.Item](sizes []int, partition bool, partitionSize int) Dataset {
+func New[T btree.Item](sizes []int64, partition bool, partitionSize int) Dataset {
 	if partition {
-		partitions := make([][]int, 0)
+		partitions := make([][]int64, 0)
 		for base := 0; base < len(sizes); base += partitionSize {
 			partitions = append(partitions, sizes[base:base+partitionSize])
 		}
@@ -65,7 +65,7 @@ type ShardedDataset[T btree.Item] struct {
 }
 
 // NewShardedDataset creates a new sharded dataset with the given argument.
-func NewShardedDataset[T btree.Item](sizes []int) *ShardedDataset[T] {
+func NewShardedDataset[T btree.Item](sizes []int64) *ShardedDataset[T] {
 	// We use the default degree for the nodes to fit on a single memory page.
 	dataset := &ShardedDataset[T]{
 		items:      btree.New[T](0),
@@ -73,7 +73,7 @@ func NewShardedDataset[T btree.Item](sizes []int) *ShardedDataset[T] {
 	}
 
 	for index, size := range sizes {
-		if _, found := dataset.items.ReplaceOrInsert(btree.NewItem[T](index, size)); found {
+		if _, found := dataset.items.ReplaceOrInsert(btree.NewItem[T](index, int(size))); found {
 			panic("insert found item")
 		}
 	}
@@ -155,7 +155,7 @@ type PartitionedDataset[T btree.Item] struct {
 }
 
 // NewPartitionedDataset creates a new partitioned dataset with the given argument.
-func NewPartitionedDataset[T btree.Item](sizes [][]int) *PartitionedDataset[T] {
+func NewPartitionedDataset[T btree.Item](sizes [][]int64) *PartitionedDataset[T] {
 	dataset := &PartitionedDataset[T]{
 		partitions:  make([]*btree.BTree[T], len(sizes)),
 		recycleBins: make([]*btree.BTree[T], len(sizes)),
@@ -169,7 +169,7 @@ func NewPartitionedDataset[T btree.Item](sizes [][]int) *PartitionedDataset[T] {
 		dataset.partitions[rank] = btree.New[T](0)
 		dataset.recycleBins[rank] = btree.New[T](0)
 		for index, size := range partition {
-			if _, found := dataset.partitions[rank].ReplaceOrInsert(btree.NewItem[T](base+index, size)); found {
+			if _, found := dataset.partitions[rank].ReplaceOrInsert(btree.NewItem[T](base+index, int(size))); found {
 				panic("insert found item")
 			}
 		}
